Updated HardwareHealth component created as functional based component health details are showed in popup instead in tooltip optimized performance by not rerendering the component if data is unchanged
